requirements.txt
streamlit>=1.41.0
Flask>=3.0.0
gunicorn>=21.2.0
opencv-python-headless>=4.10.0.84
yt-dlp>=2025.1.0
Pillow>=10.2.0
